# specs/main/contracts/api.yaml
openapi: "3.1.0"
info:
  title: Planche.bg — AI Interior Design API
  version: "1.0.0"
  description: |
    Backend API for the Planche.bg AI Interior Design platform.
    Handles sketch analysis, furniture matching, render generation, and product search.
    All prices in EUR. Product names may be in Bulgarian (Cyrillic).

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.planche.bg
    description: Production

paths:
  /api/analyze:
    post:
      summary: Analyze a room sketch
      description: |
        Upload a sketch image (photo of hand-drawn floor plan).
        Returns structured room analysis: type, dimensions, windows, doors, usable walls.
        Bulgarian labels (хол, спалня) are detected and translated to English.
      operationId: analyzeSketch
      tags: [Design]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sketch]
              properties:
                sketch:
                  type: string
                  format: binary
                  description: Sketch image (JPEG, PNG, WebP)
      responses:
        "200":
          description: Room analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomAnalysis"
        "400":
          description: Invalid image or unreadable sketch
        "500":
          description: Gemini Vision API error

  /api/furnish:
    post:
      summary: Generate a furnished room design
      description: |
        Upload sketch + set budget/style/tier → get photorealistic render with real products.
        This is the main endpoint: sketch → analyze → match products → generate render.
        Returns render URL + product cards with buy links.
      operationId: furnishRoom
      tags: [Design]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sketch, budget]
              properties:
                sketch:
                  type: string
                  format: binary
                  description: Sketch image (JPEG, PNG, WebP)
                room_id:
                  type: string
                  default: "room_1"
                  description: Room ID from analysis (for multi-room sketches)
                budget:
                  type: number
                  format: float
                  description: Budget in EUR
                  minimum: 100
                  maximum: 50000
                tier:
                  type: string
                  enum: [budget, standard, premium, luxury]
                  default: standard
                  description: |
                    Budget tier:
                    - budget (Бюджетен): €30-60/m², JYSK/IKEA/eMAG
                    - standard (Стандартен): €60-150/m², Videnov/IKEA/AIKO
                    - premium (Премиум): €150-350/m², Wayfair/IKEA/Videnov
                    - luxury (Луксозен): €350-1000/m², Wayfair
                style:
                  type: string
                  enum: [modern, scandinavian, industrial, classic, minimalist, mid-century, bohemian, rustic]
                  default: modern
      responses:
        "200":
          description: Generated design with render and product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignResult"
        "400":
          description: Invalid parameters
        "422":
          description: No matching products found for constraints
        "500":
          description: Generation error

  /api/swap:
    post:
      summary: Swap a product in an existing design
      description: |
        Replace one furniture piece in a design. If no product_id given, returns alternatives.
        If product_id given, triggers re-render with the new product's photo.
      operationId: swapProduct
      tags: [Design]
      parameters:
        - name: design_id
          in: query
          required: true
          schema:
            type: string
        - name: slot
          in: query
          required: true
          schema:
            type: string
          description: "Furniture slot to swap (e.g., 'sofa', 'coffee_table', 'bed')"
        - name: product_id
          in: query
          required: false
          schema:
            type: string
          description: "Product ID to swap in. If omitted, returns alternatives."
        - name: max_price
          in: query
          required: false
          schema:
            type: number
            format: float
      responses:
        "200":
          description: Alternatives list or re-rendered design
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SwapAlternatives"
                  - $ref: "#/components/schemas/SwapResult"
        "404":
          description: Design or product not found

  /api/products/search:
    get:
      summary: Search furniture products
      description: Search and filter the product database. Returns products with images and buy links.
      operationId: searchProducts
      tags: [Products]
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: "Text search (name, description). Supports Bulgarian text."
        - name: category
          in: query
          schema:
            type: string
            enum: [sofa, table, chair, bed, wardrobe, cabinet, desk, lamp, rug, shelf, mirror, nightstand, dresser]
        - name: room
          in: query
          schema:
            type: string
            enum: [living room, bedroom, kitchen, bathroom, office, kids room, dining room]
        - name: style
          in: query
          schema:
            type: string
        - name: min_price
          in: query
          schema:
            type: number
            format: float
        - name: max_price
          in: query
          schema:
            type: number
            format: float
        - name: source
          in: query
          schema:
            type: string
          description: "Source domain filter (e.g., 'videnov.bg')"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductSummary"

  /api/stats:
    get:
      summary: Database statistics
      description: Returns counts of total products, enriched products, and usable products.
      operationId: getStats
      tags: [Products]
      responses:
        "200":
          description: Database statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseStats"

  /api/tiers:
    get:
      summary: Get budget tier definitions
      description: Returns the 4 budget tiers with labels (Bulgarian + English), price ranges, and descriptions.
      operationId: getTiers
      tags: [Configuration]
      responses:
        "200":
          description: Budget tier definitions
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/BudgetTier"

components:
  schemas:
    RoomAnalysis:
      type: object
      properties:
        rooms:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "room_1"
              type:
                type: string
                example: "living room"
              estimated_width_cm:
                type: number
                example: 450
              estimated_depth_cm:
                type: number
                example: 380
              estimated_area_sqm:
                type: number
                example: 17.1
              features:
                type: object
                properties:
                  windows:
                    type: array
                    items:
                      type: object
                  doors:
                    type: array
                    items:
                      type: object
              usable_walls:
                type: array
                items:
                  type: object
                  properties:
                    wall:
                      type: string
                    free_length_cm:
                      type: number
                    suitable_for:
                      type: array
                      items:
                        type: string
              furniture_zones:
                type: array
                items:
                  type: object
        overall:
          type: object
          properties:
            total_rooms:
              type: integer
            total_area_sqm:
              type: number
            detected_labels:
              type: array
              items:
                type: string
              description: "Original Bulgarian labels detected in sketch"

    DesignResult:
      type: object
      properties:
        design_id:
          type: string
          example: "a1b2c3d4e5f6"
        render_url:
          type: string
          format: uri
          description: "Cloudflare R2 public URL for the generated render"
        sketch_url:
          type: string
          format: uri
        products:
          type: array
          items:
            $ref: "#/components/schemas/BuyLink"
        budget_spent:
          type: number
          format: float
          example: 1847.50
        budget_remaining:
          type: number
          format: float
          example: 152.50

    BuyLink:
      type: object
      properties:
        name:
          type: string
          description: "Product name (may be Bulgarian)"
          example: "Диван ъглов Марио"
        price:
          type: number
          format: float
          example: 549.99
        currency:
          type: string
          example: "EUR"
        url:
          type: string
          format: uri
          description: "Link to product page on source store"
        source:
          type: string
          example: "videnov.bg"
        image_url:
          type: string
          format: uri
          description: "Product image from R2"

    ProductSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: float
        currency:
          type: string
        category:
          type: string
        room_type:
          type: string
        style:
          type: string
        r2_main_image_key:
          type: string
        width_cm:
          type: number
        height_cm:
          type: number
        depth_cm:
          type: number
        rating:
          type: number
        source_domain:
          type: string
        product_url:
          type: string
          format: uri
        in_stock:
          type: boolean

    SwapAlternatives:
      type: object
      properties:
        alternatives:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              price:
                type: number
              image_url:
                type: string
                format: uri
              visual_description:
                type: string

    SwapResult:
      type: object
      properties:
        design_id:
          type: string
        render_url:
          type: string
          format: uri
        swapped_slot:
          type: string
        new_product:
          type: object
          properties:
            name:
              type: string
            price:
              type: number
            image_url:
              type: string
              format: uri

    DatabaseStats:
      type: object
      properties:
        total_products:
          type: integer
        vision_enriched:
          type: integer
        usable_for_render:
          type: integer
        enrichment_pct:
          type: number

    BudgetTier:
      type: object
      properties:
        label:
          type: string
          description: "Bilingual label"
          example: "Стандартен / Standard"
        description:
          type: string
        per_sqm_eur:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
          example: [60, 150]
        preferred_sources:
          type: array
          items:
            type: string
        max_single_item_pct:
          type: number
          example: 0.30